rename files to keep good casing
