+++ src/browser-context.js
